Refactor file structure, update imports, and add new tests
